third_party/WebKit/LayoutTests/platform/mac-retina/external/wpt/IndexedDB/idbobjectstore-rename-errors-expected.txt

This is a testharness.js-based test.
FAIL IndexedDB deleted object store rename throws assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename throws in a readonly transaction assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename throws in a readwrite transaction assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename throws in an inactive transaction assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename to the name of another store throws assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename handles exceptions when stringifying names assert_equals: Expected success event, but got error event instead expected "success" but got "error"
Harness: the test ran to completion.
